Revert "Rename api.apis to graph_reader.artifact"
This reverts commit 0cff5d4ed486058228601cd8b69434cd60bd6e41.
