+++ lineapy/api/apis.py
@@ -1,5 +1,5 @@
 """
-User exposed objects through the :mod:`lineapy.api`.
+User exposed objects through the :mod:`lineapy.apis`.
 """
 from __future__ import annotations
 
@@ -188,7 +188,7 @@ class LineaArtifact:
 class LineaArtifactStore:
     """LineaArtifactStore
 
-    A simple way to access meta data about artifacts in Linea
+    A simple way to access meta data about artifacts in Linea.
     """
 
     """
