+++ tests/test_api.py
@@ -1,4 +1,4 @@
-from lineapy.graph_reader.artifact import LineaArtifact
+from lineapy.api.apis import LineaArtifact
 
 
 def test_execute_slice(execute):
